Added the adictional dependecies line to client configuration

+++ Breakout/Server/ThreadManager.cpp
@@ -329,7 +329,7 @@ DWORD WINAPI RemoteConnectionHandler(LPVOID args) {
 	}
 
 	CONTINUE = false;
-	WaitForMultipleObjects(clientThreads.size(), clientThreads.data(), TRUE, INFINITE);
+	WaitForMultipleObjects((DWORD) clientThreads.size(), clientThreads.data(), TRUE, INFINITE);
 
 	return 0;
 }
